Fix the mimatch file name for tests of filter operators
