+++ packages/api_tests/__tests__/undefinable/filter/filter.test.mjs
@@ -1,12 +1,12 @@
 import test from 'ava';
 
-import { filterWithEnsureTypeForUndefinable } from 'option-t/Undefinable/filter';
+import { filterForUndefinable } from 'option-t/Undefinable/filter';
 
 test('input is T, and predicate return true', (t) => {
     t.plan(2);
 
     const INPUT = Math.random();
-    const actual = filterWithEnsureTypeForUndefinable(INPUT, (inner) => {
+    const actual = filterForUndefinable(INPUT, (inner) => {
         t.is(inner, INPUT);
         return true;
     });
@@ -17,7 +17,7 @@ test('input is T, and predicate return false', (t) => {
     t.plan(2);
 
     const INPUT = Math.random();
-    const actual = filterWithEnsureTypeForUndefinable(INPUT, (inner) => {
+    const actual = filterForUndefinable(INPUT, (inner) => {
         t.is(inner, INPUT);
         return false;
     });
@@ -27,7 +27,7 @@ test('input is T, and predicate return false', (t) => {
 test('input is undefined', (t) => {
     t.plan(1);
 
-    const actual = filterWithEnsureTypeForUndefinable(undefined, (_inner) => {
+    const actual = filterForUndefinable(undefined, (_inner) => {
         t.fail();
         return true;
     });
