+++ packages/api_tests/__tests__/undefinable/filter/filter_with_ensure_type.test.mjs
@@ -1,12 +1,12 @@
 import test from 'ava';
 
-import { filterForUndefinable } from 'option-t/Undefinable/filter';
+import { filterWithEnsureTypeForUndefinable } from 'option-t/Undefinable/filter';
 
 test('input is T, and predicate return true', (t) => {
     t.plan(2);
 
     const INPUT = Math.random();
-    const actual = filterForUndefinable(INPUT, (inner) => {
+    const actual = filterWithEnsureTypeForUndefinable(INPUT, (inner) => {
         t.is(inner, INPUT);
         return true;
     });
@@ -17,7 +17,7 @@ test('input is T, and predicate return false', (t) => {
     t.plan(2);
 
     const INPUT = Math.random();
-    const actual = filterForUndefinable(INPUT, (inner) => {
+    const actual = filterWithEnsureTypeForUndefinable(INPUT, (inner) => {
         t.is(inner, INPUT);
         return false;
     });
@@ -27,7 +27,7 @@ test('input is T, and predicate return false', (t) => {
 test('input is undefined', (t) => {
     t.plan(1);
 
-    const actual = filterForUndefinable(undefined, (_inner) => {
+    const actual = filterWithEnsureTypeForUndefinable(undefined, (_inner) => {
         t.fail();
         return true;
     });
